Restructure connection state machine files
