+++ libvcx/src/utils/libindy/anoncreds.rs
@@ -13,7 +13,6 @@ use utils::libindy::cache::{clear_rev_reg_delta_cache, get_rev_reg_delta_cache, 
 use utils::libindy::ledger::*;
 use utils::libindy::payments::{pay_for_txn, PaymentTxn};
 use utils::mockdata::mock_settings::get_mock_creds_retrieved_for_proof_request;
-use v3::handlers::issuance::utils::encode_attributes;
 use v3::messages::proof_presentation::presentation_request::{PresentationRequestData, PresentationRequest};
 use utils::mockdata::mockdata_proof;
 
@@ -622,6 +621,7 @@ pub mod tests {
     use utils::get_temp_dir_path;
 
     use super::*;
+    use v3::handlers::issuance::issuer::utils::encode_attributes;
 
     extern crate serde_json;
     extern crate rand;
