+++ libvcx/src/v3/handlers/connection/agent_info.rs
