+++ libvcx/src/v3/handlers/connection/states/requested.rs
@@ -1,0 +1,60 @@
+use error::prelude::*;
+use v3::handlers::connection::agent_info::AgentInfo;
+use v3::handlers::connection::states::complete::CompleteState;
+use v3::handlers::connection::states::null::NullState;
+use v3::messages::ack::Ack;
+use v3::messages::connection::did_doc::DidDoc;
+use v3::messages::connection::problem_report::{ProblemCode, ProblemReport};
+use v3::messages::connection::request::Request;
+use v3::messages::connection::response::{Response, SignedResponse};
+use v3::messages::trust_ping::ping::Ping;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct RequestedState {
+    pub request: Request,
+    pub did_doc: DidDoc,
+}
+
+
+impl From<(RequestedState, ProblemReport)> for NullState {
+    fn from((_state, _error): (RequestedState, ProblemReport)) -> NullState {
+        trace!("DidExchangeStateSM: transit state from RequestedState to NullState");
+        NullState {}
+    }
+}
+
+impl From<(RequestedState, Response)> for CompleteState {
+    fn from((_state, response): (RequestedState, Response)) -> CompleteState {
+        trace!("DidExchangeStateSM: transit state from RequestedState to RespondedState");
+        CompleteState { did_doc: response.connection.did_doc, protocols: None }
+    }
+}
+
+impl RequestedState {
+    pub fn handle_connection_response(&self, response: SignedResponse, agent_info: &AgentInfo) -> VcxResult<Response> {
+        trace!("RequestedState:handle_connection_response >>> response: {:?}, agent_info: {:?}", response, agent_info);
+
+        let remote_vk: String = self.did_doc.recipient_keys().get(0).cloned()
+            .ok_or(VcxError::from_msg(VcxErrorKind::InvalidState, "Cannot handle Response: Remote Verkey not found"))?;
+
+        let response: Response = response.decode(&remote_vk)?;
+
+        if !response.from_thread(&self.request.id.0) {
+            return Err(VcxError::from_msg(VcxErrorKind::InvalidJson, format!("Cannot handle Response: thread id does not match: {:?}", response.thread)));
+        }
+
+        let message = if response.please_ack.is_some() {
+            Ack::create()
+                .set_thread_id(&response.thread.thid.clone().unwrap_or_default())
+                .to_a2a_message()
+        } else {
+            Ping::create()
+                .set_thread_id(response.thread.thid.clone().unwrap_or_default())
+                .to_a2a_message()
+        };
+
+        agent_info.send_message(&message, &response.connection.did_doc)?;
+
+        Ok(response)
+    }
+}
