+++ libvcx/src/v3/handlers/proof_presentation/prover/states/presentation_prepared.rs
@@ -1,0 +1,34 @@
+use v3::handlers::proof_presentation::prover::states::finished::FinishedState;
+use v3::handlers::proof_presentation::prover::states::presentation_sent::PresentationSentState;
+use v3::messages::proof_presentation::presentation::Presentation;
+use v3::messages::proof_presentation::presentation_request::PresentationRequest;
+use v3::messages::status::Status;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct PresentationPreparedState {
+    pub presentation_request: PresentationRequest,
+    pub presentation: Presentation,
+}
+
+impl From<(PresentationPreparedState, u32)> for PresentationSentState {
+    fn from((state, connection_handle): (PresentationPreparedState, u32)) -> Self {
+        trace!("transit state from PresentationPreparedState to PresentationSentState");
+        PresentationSentState {
+            presentation_request: state.presentation_request,
+            presentation: state.presentation,
+            connection_handle,
+        }
+    }
+}
+
+impl From<PresentationPreparedState> for FinishedState {
+    fn from(state: PresentationPreparedState) -> Self {
+        trace!("transit state from PresentationPreparedState to FinishedState");
+        FinishedState {
+            connection_handle: 0,
+            presentation_request: state.presentation_request,
+            presentation: Default::default(),
+            status: Status::Declined,
+        }
+    }
+}
