+++ libvcx/src/v3/handlers/proof_presentation/verifier/state_machine.rs
@@ -12,6 +12,9 @@ use v3::messages::proof_presentation::presentation_ack::PresentationAck;
 use v3::messages::proof_presentation::presentation_request::{PresentationRequest, PresentationRequestData};
 use v3::messages::status::Status;
 use proof_utils::validate_indy_proof;
+use v3::handlers::proof_presentation::verifier::states::initial::InitialState;
+use v3::handlers::proof_presentation::verifier::states::presentation_request_sent::PresentationRequestSentState;
+use v3::handlers::proof_presentation::verifier::states::finished::FinishedState;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct VerifierSM {
@@ -32,82 +35,11 @@ pub enum VerifierState {
     Finished(FinishedState),
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct InitialState {
-    presentation_request_data: PresentationRequestData
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct PresentationRequestSentState {
-    connection_handle: u32,
-    presentation_request: PresentationRequest,
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum RevocationStatus {
     Revoked,
     NonRevoked,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FinishedState {
-    connection_handle: u32,
-    presentation_request: PresentationRequest,
-    presentation: Option<Presentation>,
-    status: Status,
-    revocation_status: Option<RevocationStatus>,
-}
-
-impl From<(InitialState, PresentationRequest, u32)> for PresentationRequestSentState {
-    fn from((_state, presentation_request, connection_handle): (InitialState, PresentationRequest, u32)) -> Self {
-        trace!("transit state from InitialState to PresentationRequestSentState");
-        PresentationRequestSentState { connection_handle, presentation_request }
-    }
-}
-
-impl From<(PresentationRequestSentState, Presentation, RevocationStatus)> for FinishedState {
-    fn from((state, presentation, was_revoked): (PresentationRequestSentState, Presentation, RevocationStatus)) -> Self {
-        trace!("transit state from PresentationRequestSentState to FinishedState");
-        FinishedState {
-            connection_handle: state.connection_handle,
-            presentation_request: state.presentation_request,
-            presentation: Some(presentation),
-            status: Status::Success,
-            revocation_status: Some(was_revoked),
-        }
-    }
-}
-
-impl From<(PresentationRequestSentState, ProblemReport)> for FinishedState {
-    fn from((state, problem_report): (PresentationRequestSentState, ProblemReport)) -> Self {
-        trace!("transit state from PresentationRequestSentState to FinishedState");
-        FinishedState {
-            connection_handle: state.connection_handle,
-            presentation_request: state.presentation_request,
-            presentation: None,
-            status: Status::Failed(problem_report),
-            revocation_status: None,
-        }
-    }
-}
-
-
-impl PresentationRequestSentState {
-    fn verify_presentation(&self, presentation: &Presentation) -> VcxResult<()> {
-        let valid = validate_indy_proof(&presentation.presentations_attach.content()?,
-                                               &self.presentation_request.request_presentations_attach.content()?)?;
-
-        if !valid {
-            return Err(VcxError::from_msg(VcxErrorKind::InvalidProof, "Presentation verification failed"));
-        }
-
-        if presentation.please_ack.is_some() {
-            let ack = PresentationAck::create().set_thread_id(&self.presentation_request.id.0);
-            connection::send_message(self.connection_handle, A2AMessage::PresentationAck(ack))?;
-        }
-
-        Ok(())
-    }
 }
 
 impl VerifierSM {
