+++ libvcx/src/v3/handlers/proof_presentation/verifier/verifier.rs
@@ -1,7 +1,7 @@
 use ::{connection};
 use error::prelude::*;
 use v3::handlers::proof_presentation::verifier::messages::VerifierMessages;
-use v3::handlers::proof_presentation::verifier::states::VerifierSM;
+use v3::handlers::proof_presentation::verifier::state_machine::VerifierSM;
 use v3::messages::a2a::A2AMessage;
 use v3::messages::proof_presentation::presentation::Presentation;
 use v3::messages::proof_presentation::presentation_request::*;
